Split error module tests
